requirement.txt
uvicorn
fastapi
couchbase 
passlib
python-multipart
jinja2
aiofiles
pyjwt